Adjust function signature so caching has effect.
We wrapped that atomic `collect(release)` method in a cache to make
things fast, but it never worked (it seems).  We *modify* that `release`
dict on every pass, so its different every time which then always
bypasses the cache.

In this change, we stop passing that whole dict of releases through,
since we only need a few small values.  We pass those instead now.
Caching works, and the site build time is faster.