Update website to query f26 atomic
The current code that collects data will blow up if it can't find
a fedmsg related to a two week atomic release which means the whole
website can't be built. This has forced us to leave the variables
at f25 for now.

We'll merge this commit right after F26AH is released on tuesday so
that the website can get the data it needs from the F26 fedmsg that
was just sent out.