Do not use datagrepper's contains, but filter locally, and submit a delta
This moves compute time and memory usage from the datagrepper server to the system running
the fedimg_vars scripts.
Without delta, we were asking the datagrepper database to check every message since
datanommer was started for the first time.
With contains, we ask postgres to perform a heavy string comparison operation.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>