[getfedora.org] add item to news section for updated privacy policy
This adds a new note to the news box on the getfedora.org index
page that informs users that the Privacy Policy has been updated.

The privacy policy itself dictates we do this for 30 days after a
change, so put this into both the GA and Beta news boxes, in case
we flip to Beta during the 30 Days.

Signed-off-by: Ryan Lerch <rlerch@redhat.com>